Change name of CardsContainer and move display card projects from search to cardsContainer component
