+++ app/src/components/ProjectList/CardsContainer.css
@@ -1,4 +1,5 @@
 .Container-layout{
+    background: white;
     display: -ms-flexbox;
     padding: 1.2em;
     overflow: hidden;
